==> Procesos/5_Limpieza_de_datos_y_agregaciones.ipynb <==
# Limpieza de datos y agregaciones

### OBJETIVO

- Limpiar nuestros datasets de `NaNs`.
- Reindexar si es necesario
- Renombrar columnas si es necesario
- Experimentar la aplicación de agregaciones para explorar nuestro dataset

## Carga general de datos

Para nuestra limpieza de datos, solo utilizaremos las siguientes librerias:

import pandas as pd
import numpy as np

Vamos a realizar lo siguiente:

- Cargar los documentos que limpiaremos, con el objetivo de dejar un par de archivos `csv` al final
- Concentrar los datos por tipo de PM

df_PM10_2019 = pd.read_csv('../Datasets/PM10/2019PM10.csv') #Data PM 10 - 2019
df_PM10_2020 = pd.read_csv('../Datasets/PM10/2020PM10.csv') #Data PM 10 - 2020
df_PM25_2019 = pd.read_csv('../Datasets/PM2.5/2019PM25.csv') #Data PM 2.5 - 2019
df_PM25_2020 = pd.read_csv('../Datasets/PM2.5/2020PM25.csv') #Data PM 2.5 - 2020

Después de cargar nuestros archivos `csv`, los concatenamos para tener dos Dataframes (df), con los cuales vamos a realizar nuestra depuración de datos.

df_PM10_total = pd.concat([df_PM10_2019 , df_PM10_2020], axis=0)
df_PM25_total = pd.concat([df_PM25_2019 , df_PM25_2020], axis=0)

## Depuración de archivos $PM_{10}$

Comenzamos depurando la información de $PM_{10}$  
De acuerdo a las especificaciones RAMA, los datos nulos se identifican con la etiqueta -99, por lo que, convertimos los valores `-99` en `NaN`. 

df_PM10_total = df_PM10_total.replace(-99, np.nan)

Haciendo una rápida validación de nuestro df que concentra los datos de $PM_{10}$, vemos que contiene **17,544** `filas` y **29** `columnas`.

df_PM10_total.shape

Ahora hacemos una revisión de los datos nulos que contiene nuestro df.   
Podemos ver, que todas las filas y las columnas de las estaciones cuentan con al menos un valor `NaN`

# Filas con NaN
df_PM10_total.isna().any(1).sum()

# Columnas con NaN
df_PM10_total.isna().any()

Revisando la parte proporcional de `NaNs` por cada columna, observamos que la mayoría de nuestras estaciones cuentan con un alto porcentaje de datos nulos.
Y que la media de nuestras medias del df corresponde al **39.76%** de `NaN`. Una cifra bastante alta, por cierto.

df_PM10_total.isna().mean()

df_PM10_total.isna().mean().mean() * 100

Considerando lo anterior, vamos a eliminar aquellas columnas que contengan al menos el **55%** de valores `NaNs`  
Con esto, nos estamos quedando ahora con **23** columnas en total.

# Se eliminan con el 55% de NaN
df_PM10_depurado = df_PM10_total.dropna(thresh=len(df_PM10_total)*0.45, axis=1)

df_PM10_depurado.shape

A pesar de que eliminamos columnas con un alto porcentaje de nulos, al validar nuestro df nuevamente, nos percatamos que todas nuestras filas tienen al menos un valor `NaN`, y ninguna fila con todos los valores `NaN`. Por lo tanto, no podemos eliminar las filas, debido a que nos estaríamos quedando sin datos para procesar.  
Tampoco es posible, por esta estructura que tienen nuestros df, reemplazar estos valores `NaN` con 0, debido a que la naturaleza de nuestros análisis requieren obtener medias, y este valor nos estaría dando una tendencia errónea en nuestros resultados.

df_PM10_depurado[df_PM10_depurado.isna().any(axis=1)].shape[0]

df_PM10_depurado.dropna(axis=0, how='all').shape

Realizamos un ejemplo de agregado (mean), que se puede aplicar a nuestro df depurado

df_PM10_depurado.mean()

Por último, guardamos nuestro df depurado en un nuevo `csv` de $PM_{10}$

#Guardar csv
df_PM10_depurado.to_csv('../Datasets/PM10/PM10_depurado.csv')

## Depuración de archivos $PM_{2.5}$

Ahora depuramos la información de $PM_{2.5}$  
De acuerdo a las especificaciones RAMA, los datos nulos se identifican con la etiqueta -99, por lo que, convertiremos los valores `-99` en `NaN`. 

df_PM25_total = df_PM25_total.replace(-99, np.nan)

Haciendo una rápida validación de nuestro df que concentra los datos de $PM_{2.5}$, vemos que contiene **16,800** `filas` y **26** `columnas`.

#Obtenemos las columnas
df_PM25_total.shape

Aplicamos la revisión de los datos nulos que contiene nuestro df.  
Podemos ver que en este df, también todas las filas y las columnas de las estaciones cuentan con al menos un valor `NaN`

# Filas con NaN
df_PM25_total.isna().any(1).sum()

# Columnas con NaN
df_PM25_total.isna().any()

Revisando la parte proporcional de `NaNs` por cada columna, observamos que la mayoría de nuestras estaciones también cuentan con un alto porcentaje de datos nulos.
En este caso, la media de nuestras medias del df corresponde al **39.69%** de `NaN`. Una cifra igual de alta, por cierto.

df_PM25_total.isna().mean()

df_PM25_total.isna().mean().mean() * 100

Al igual que en nuestro df anterior, vamos a eliminar aquellas columnas que contengan al menos el **55%** de valores `NaNs`  
Con esto, nos estamos quedando ahora con **22** columnas en total.

# Se eliminan con el 55% de NaN

df_PM25_depurado = df_PM25_total.dropna(thresh=len(df_PM25_total)*0.45, axis=1)

df_PM25_depurado.shape

En este caso, nos percatamos que salvo 2 filas, en este df las filas contienen al menos un valor `NaN`, y ninguna fila con todos los valores `NaN`. Por lo tanto, no podemos eliminar las filas, debido a que nos estaríamos quedando sin datos para procesar.  
Tampoco es posible, por esta estructura que tienen nuestros df, reemplazar estos valores `NaN` con 0, debido a que la naturaleza de nuestros análisis requieren obtener medias, y este valor nos estaría dando una tendencia errónea en nuestros resultado.

df_PM25_depurado[df_PM25_depurado.isna().any(axis=1)].shape

Realizamos un ejemplo de agregado (mean), que se puede aplicar a nuestro df depurado

df_PM25_depurado.mean()

Por último, guardamos nuestro df depurado en un nuevo `csv` de $PM_{2.5}$

#Guardar csv
df_PM25_depurado.to_csv('../Datasets/PM2.5/PM25_depurado.csv')

## Depuración de cat_estacion

El siguiente archivo a depurar, es el catálogo de estaciones y zonas.  
Cargamos nuestro archivo, y validamos 

df_cat_estacion = pd.read_csv('../Datasets/cat_estacion.csv') #Data catálogo de estaciones

Haciendo una rápida validación de nuestro df con el catálogo de estaciones y zonas, vemos que contiene **69** `filas` y **8** `columnas`.

df_cat_estacion.shape

Verificamos, cuántas filas y columnas cuentan con datos `NaN`  
Confirmamos que en este df no tenemos tantos nulos, por lo que borraremos laa columnas que contengan datos `NaN` y que no hacen inferencia en nuestro análisis

df_cat_estacion.isna().any(1).sum()

df_cat_estacion.isna().any()

df_cat_estacion.isna().mean()

df_cat_estacion = df_cat_estacion.drop(['longitud','latitud','alt','obs_estac','id_station'], axis=1)

Al final, después de depurar, solo nos quedamos con **3** columnas y ningún valor `NaN`

df_cat_estacion.isna().mean()

Guardamos nuestro nuevo catalogo de estaciones depurado en otro `csv`

df_cat_estacion.to_csv('../Datasets/cat_estacion_depurado.csv')
