==> Procesos/6_1_Transformacion_PM10.ipynb <==
# Promedio Diario $PM_{10}$

Para nuestra transformación de datos, utilizaremos las siguientes librerias:



import pandas as pd
import numpy as np

- Cargando dataset a un DataFrame de pandas

- Posteriormente normalizando la columna `Fecha` a tipo datetime

#Cargando archivos
df = pd.read_csv('../Datasets/PM10/PM10_depurado.csv')

# Normalizando fechas
df['FECHA'] = pd.to_datetime(df['FECHA'], unit='ns')
df.head()

Posteriormente, se filtran los datos por los meses de interés, que son marzo, abril y mayo. Primeramente se crea el rango de fechas de interés y luego éste se utiliza para filtrar el dataset

date_rng = pd.date_range(start='2019/03/01', end='2019/05/31', freq='d')#Rango en 2019
date_rng2= pd.date_range(start='2020/03/01', end='2020/05/31', freq='d') #Rango en 2020
date_rng= date_rng.append(date_rng2) #unión de rangos

df=df[df["FECHA"].isin(date_rng)]
df.head()

- Reacomodo del DataFrame , donde ahora tendremos una columna para las estaciones y sus correspondientes datos de medición
- Eliminando los datos `NaN`
- Comprobando que no queden datos `NaN` en el DataFrame

# Extraer nombre de columnas
cols = df.columns.tolist()

# Reacomodando dataframe
df = pd.melt(df, id_vars=['FECHA', 'HORA'], value_vars=cols[3:], var_name='station', value_name='measurement')

# df = df[~df['measurement'].isna()] otra forma de eliminar nan's
df = df.dropna(axis=0, how='any').reset_index(drop=True)

print("No. de columnas que contienen valores nulos ")
print(len(df.columns[df.isna().any()]))

print("No. de columnas que no contienen valores nulos")
print(len(df.columns[df.notna().all()]))

print("No total de columnas en el dataframe")
print(len(df.columns))
df.head()

- Ordenando el dataframe segun la fecha y la estación

# ordenando el dataframe por fecha y nombre de estacion
df = df.sort_values(['FECHA', 'station']).reset_index(drop=True)
df

Ahora se busca eliminar, según el día y la estación, aquellas filas que tengan mediciones horarias menores a las requeridas por la normativa (18 mediciones). Por ejemplo, si la estación ACO en el 01 de enero de 2019 registró solamente 15 mediciones horarias, todas las mediciones de ese día para la estación ACO no será tomadas en cuenta para los posteriores análisis. 

Primero se debe elaborar una función que devuelva el dataframe que contenga solamente los registros válidos

def dia_estacion_valido(dataframe):
  #Contar las mediciones horarias por día y estación
  filtro= dataframe.groupby(['FECHA', 'station'],as_index=False).agg(N_h=('HORA','count'))
  #Filtrar aquellas que no cumplen normativa
  filtro= filtro.query('N_h < 17')
  #Conservar las columnas de fecha y estación
  filtro=filtro.loc[:,['FECHA','station']]
  #Eliminar del dataframe original las mediciones que no cumplen normativa
  filtro=pd.merge(dataframe,filtro, how='outer', indicator=True)
  filtro=filtro[filtro['_merge'] == 'left_only']
  return filtro


Con la función, se filtra las mediciones por día y por estación son válidas para los siguientes análisis.


df_filtrado= dia_estacion_valido(df)
df_filtrado = df_filtrado.drop(columns=['_merge'])
df_filtrado=df_filtrado.reset_index()
df_filtrado = df_filtrado.drop(columns=['index'])
df_filtrado


Una vez con el dataframe inicial limpio, se procede a relacionar cada estación con su respectiva zona de activación. Primeramente se leen los datos

# Cargando dataset de zonas
df_zonas = pd.read_csv('../Datasets/cat_estacion_depurado.csv')

# Eliminando columna 'Unnamed: 0'
df_zonas = df_zonas.drop(columns=['Unnamed: 0'])
# Remplazando los - por NaN
df_zonas = df_zonas.replace('-', np.nan)
# Eliminando lo NaN's
df_zonas = df_zonas.dropna(axis=0, how='any').reset_index(drop=True)
# Renombrando la columna
df_zonas = df_zonas.rename(columns={'cve_estac': 'station'})
df_zonas.head()

- Uniendo los datos con las zonas correspondiente


df_filtrado = df_filtrado.merge(df_zonas[['station', 'Zona']], on='station')
df_filtrado.head()

Con los datos válidos para realizar las operaciones siguientes y ya con la relación de cada estación con su respectiva zona de activación, se procede a calcular el promedio diario de concentración y el índice de calidad del aire en cada zona y en cada día

### Calculando el promedio diario
Para el cálculo del promedio diario de $PM_{10}$ se utilizó la metodología descrita en la `NOM-025-SSA1-2014`, siguiendo la siguiente ecuación:

$$\bar{x}= \frac{1}{n} \displaystyle\sum_{i=1}^n x_i$$
$\bar{x}:$ promedio de 24 horas, 

$n:$ número de concentraciones horarias válidas

$x_i:$ concentraciones horarias válidas

- Camo datos extra obteniendo los valores mínimos y máximos


# Calculando promedio diario, contando numero de horas usadas, obteniendo el máximo 
#y mínimo para las visualizaciones en el módulo siguiente
df_PM1024h = df_filtrado.groupby(['FECHA', 'Zona'], as_index=False).agg(
            PromDiario=('measurement', lambda x: round(x.mean(),0)), 
            Max = ('measurement', 'max'),
            Min = ('measurement', 'min'))
df_PM1024h.head()

### Calculando Índice de calidad del aire

Donde:
$$\text{Índice}= (k \times (C - BP_{Lo}))+I_{Lo}$$
$$k= \frac{I_{Hi}-I_{Lo}}{BP_{Hi}-BP_{Lo}}$$
 
$\text{Índice}:$ Valor del índice para el contaminante deseado.<br>
$C:$ valor redondeado para la concentración del contaminante.<br>
$k:$ constante de proporcionalidad estimada.<br>
$BP_{Hi}:$ valor del punto de corte que es mayor o igual a la concentración a evaluar.<br>
$BP_{Lo}:$ valor del punto de corte que es menor o igual a la concentración a evaluar.<br>
$I_{Hi}:$ valor del índice que corresponde al valor de $BP_{Hi}$<br>
$I_{Lo}:$ valor del índice que corresponde al valor de $BP_{Lo}$<br>


Primero se crea una función que determine el índice de calidad del aire según la metodología y los valores de la tabla descrita en el apartado Metodología para los valores de los puntos de corte

def AQI(df):
    CP = df['PromDiario']
    i = 0

    if 0 <= CP <= 40:
        k = (50-0)/(40-0)
        i = k * (CP - 0) + 0
    if 41 <= CP <= 75:
        k = (100-51)/(75-41)
        i = k * (CP - 41) + 51
    if 76 <= CP <= 214:
        k = (150-101)/(214-76)
        i = k * (CP - 76) + 101
    if 215 <= CP <= 354:
        k = (200-151)/(354-215)
        i = k * (CP - 215) + 151
    if 355 <= CP <= 424:
        k = (300-201)/(242-355)
        i = k * (CP - 355) + 201
    if 425 <= CP <= 504:
        k = (400-301)/(504-425)
        i = k * (CP - 425) + 301
    if 505 <= CP <= 604:
        k = (500-401)/(604-505)
        i = k * (CP - 505) + 401
    
    return round(i, 3)

Luego se aplica la función para cada promedio diario calculado



df_PM1024h['indice'] = df_PM1024h.apply(AQI, axis=1)
df_PM1024h.head()

Con el índice de calidad del aire, se crea otra función para determinar la categoría de calidad del aire descrita por la norma mexicana.

def AQI_categoria(df):
    indice = df['indice']
    if indice <= 50:
      clase='Buena'
    elif indice <= 100:
      clase='Regular'
    elif indice <= 150:
      clase= 'Mala'
    elif indice <= 200:
      clase = 'Muy mala'
    else: 
      clase = 'Extremadamente mala'
    
    return clase

Se aplica la función para determinar la categoría a cada promedio diario calculado.

df_PM1024h['clase'] = df_PM1024h.apply(AQI_categoria, axis=1)
df_PM1024h.head()

Se guardan los resultados en un csv

df_PM1024h.to_csv('../Datasets/PM10/PDEstacion_PM10.csv', index=False)
