==> Procesos/7_Uso_APIs.ipynb <==
# Uso de APIs

## OBJETIVO 

- Practicar el uso de APIs y peticiones HTTP.

El uso de una `API` para la obtención de datos no fue posible ya que actualmente no hay una API oficial, solo datasets que se pueden obtener el la página de Calidad de Aire de la Ciudad de México.

Con fines didácticos nosotros procedimos a realizar una pequeña API muy sencilla de los datasets obtenidos en dicha página, para poder aplicar lo aprendido en las sesiones de nuestro módulo de python.  
[Aquí describimos el procedimiento de creación](https://github.com/IrvingC48/BeduFase2-Proyecto_python/tree/main/Api)

Para realizar una consulta tenemos que visitar la siguiente dirección: http://34.73.220.231/

En la dirección anterior se documentó una serie de endpoints que sirven para obtener diferentes consultas de datos, como datos por año de una de las partículas ($PM_{10}$ o $PM_{2.5}$).


## Conexión a la API

Para conectarnos a nuestra API, utilizaremos las siguientes librerias:

import requests as res
import pandas as pd

Indicamos nuestro endpoint, y para está API no se requiere un diccionario de parámetros, ya que no solicita una Key.  
Envíamos nuestro get con el endpoint generado, y leemos el estatus de nuestro objeto de respuesta:  

Con esto, confirmamos que nuestra conexión fue exitosa

endpoint = 'http://34.73.220.231/api/dpmdiez/primer_año_all'
r = res.get(endpoint)
r.status_code

Directamente nuestra `data` es una lista de diccionarios que contiene los datos que queremos utilizar. 

Lo convertimos en un `DataFrame`, y verificamos el `head`:

json = r.json()
normalized = pd.json_normalize(json)
df = pd.DataFrame.from_dict(normalized)
df.head()

Por último, cerramos nuestra conexión a la API

r.connection.close()

Y de esta manera, aplicamos en un pequeño ejemplo el consumo de la `API` que contiene nuestros `datasets` originales
